Make Timestamp depend on Hash
This is mainly an ergonomics benefit. Since the timestamp in a scope is
often required to be `Hash`, generic code needs to specify that
explicitly. With this change it is implied by the `Timestamp` bound.

Since all existing implementations are `Hash` and the general intention
of a timestamp seems to be something that behaves mostly like an
integer, requiring `Hash` is probably not too much to ask. (Still it's
technically a breaking change.)