Add interfaces for activating across thread boundaries
Add interfaces that make it easy to send activations across thread
boundaries.

The approach is to give `Activations` an MPSC channel which it checks
for activation requests on every call to `advance`. Other threads can
acquire a handle to the send side of this object via `SyncActivations`
or `SyncActivator`, which are careful to also call unpark on the
associated worker thread for easy interop with `Worker.step_or_park`.