Improve `is_acyclic`.
The method to test acyclicity used to maintain graph structure in a hash map of vectors, which resulted in a fair amount of allocation. We already had the graph structure, but it required a bit of interpretation (constructing from internal summaries, and restricting to those with default timestamp actions). Using this data instead results in less memory churn.