Use timely's logging infrastructure to log Tracker state
Add TrackerEvent which records additons or removals
of capabilities, as well as propagation events when chanages
in implications are propagated along the internal connections
and edges of the graph.

Add DebugEvent which records the state of pointstamps,
implications, and worklist of Tracker.

Enabling loggers for these events is done the same way as for
logging::TimelyEvent's.