Add deref (with target &[Capability<T>]) and from_elem to CapabilitySet
`from_elem` allocates a `CapabilitySet` containing a single capability.
This is useful when making a `CapabilitySet` out of the default
capability provided to a new operator (as shown in the example for
`from_elem`).

`deref` lets the operator use the capabilities in the set to e.g. open a
session and send messages (again, as shown in the example for
`from_elem`).