Add PartialOrder impls (#322)
This PR adds PartialOrder implementations for Antichain and AntichainRef. The methods less_equal and less_than clash with the per-time implementations (perhaps those should only be for antichains, which each singleton time is), but the inherent methods take priority in resolution, and this doesn't appear to cause any breakage.

The intent here is to allow these types to implement Lattice in the differential crate, in the interest of tidying up a bunch of bespoke logic.