Encapsulate the operator logic in a trait
The current implementation stores functions encapsulating the logic of
an operator. In this change, we introduce a trait that describes the
behavior of the logic, compatible to the current public interface.
The hope is that later we can extend the trait to allow operators in
different stages of the computation's life cycle, while maintaining
compatibility with the current interface for operators that do not need
a more elaborate interface.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>