Make Stream generic over container type
Currently, the Stream abstraction has an opinion of the content type of
message bundles, which is that all datums are contained in vectors. With
this change, Streams are generic over the container type, allowing user
code to supply other container types if required.

This change exposes the lower-level building blocks to provide generic
operators over the container type, but many operators still require the
data to be contained by vectors.

Enables the drain function on containers implementing DrainContainer.
Due avoid lifetime complications, the trait needs to be implemented with
a known lifetime, for example using

   impl<C: Container> ... where for<'a> &'a mut C: DrainContainer

Container builder: Separate container from building a container

CoreStream: Alias core stream over vectors as `Stream` to limit API changes

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>