Implement broadcast with custom Pact
Currently, the broadcast operator duplicates the data for each peer and
uses the regular exchange infrastructure to send the duplicates to each
peer, such that each peer will receive a complete copy of the stream.

This is problematic as on the sending worker as many copies of the data
as there are peers need to exist. Also, the operator is implemented as
three operators, which needs additional copies of the data.

With this change, a specialized parallelization contract handles
distributing the data. Copies are avoided by passing reference-counted
batches of data to the peer, which means that all peers get to see the
same data, at least for process-local communication.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>