Logging: Do not flush on drop, only on last copy (#420)
The current implementation of the logger flushes when dropping the outer
logger, which contains an Rc to the inner logger. This causes excessive
flushes for clients obtaining loggers frequently. With this change, the
flush is only triggered when dropping the inner logger, i.e. when the last
copy of the Rc goes out of scope.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>