Promote non-preallocating exchange to default
I did some measurements comparing the
* current default,
* a non-preallocating exchange
* id. plus eagerly releasing exchange.

The benchmark in the chain of commits does not report a significant change
in performance for all variants. Erring on the side of caution, I'd propse
to make the second variant, i.e. not pre-allocating, but retaining if push
returned a buffer, the default exchange pact.

Currently, all non-zerocopy allocators will not return an allocation on
push, so in this case, the second and third variant are equivalent.

# Why is there no change in performance?

For single-recrod exchanges, this can be reasoned as follows. In the old
approach, we'd have an allocated buffer, where we write a single element.
On flush, the buffer would be sent to the pusher, which might or might not
return a buffer. If it doesn't, we'd allocate a new one.

In the new approach, we don't have a buffer to append to, so we'd first
allocate one. On flush, we send the buffer downstream and maybe get one
back. We don't eagerly allocate a new one if there is no buffer available.

Both approaches allocate a new buffer if the pusher doesn't/didn't return
one earlier. Only the order of allocations is changed. While in the old
approach we send data and allocate, we allocate and send data in the new
approach. On top of this, the new approach needs to pay a cost of checking
the buffer's capacity for each record on the input, but I'd hope for the
optimizer to sort this out.

The same argument seems to hold for larger batch sizes.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>