exchange: De-duplicate partitioning logic
Currently, the partitioning logic has two separate branches for
power-of-two peers using bit masking and mod for all other configurations.
The branches duplicate the logic to distribute the data to the pushers.
This change moves the common functionality into a separate function.

We force inlining of the partition function to avoid a performance penalty.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>