Specialize the container type for antichains
Antichains are currently unconditionally heap-allocated. For totally
ordered elements, an antichain will container zero or one elements, but not
more. This would allow us to avoid the heap allocation and instead store
the element inline.

To enable this, we add an associated type to `Timestamp` that defines a
container storing an antichain of the type. The container implements the
functionality to let the current `Antichain` type function without change.

The obvious downside of this change is that we now include a type bound
`T: Timestamp` on the `Antichain` type, because the elements are based on
`T`'s associated type, which infects a bunch of other places.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>