dataflow: Account for probe handles being dropped
The lifetime of a probe handle is not attached to any particular
dataflow and users are free to connect them to streams of multiple
dataflows.

This PR handles the situation in which a probe handle is connected to two
separate dataflows and one of them drops. Previously the handle would
have its frontier stuck because the old handle wouldn't provide any
updates anymore.

The solution is that each handle keeps track of its own changes and
removes them from the overall calculation on drop.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>