offer mutable access to containers through handles
This PR refactors a pattern that is pervasive to pretty much every
single timely operator into an easier to use input handle method that
also reduces the boilerplate code operator authors need to write.

The pattern is receiving a `RefOrMut` container from a handle and
immediately swapping it with a local container in order to gain mutable
access. In this repo all but one operators exhibited this pattern and
associated boilerplate.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>