Add cease to output handles
In certain situations, a handle survives longer than we would like to wait
for a flush to happen. In this cases, an explicit call to cease can help
to indicate to the rest of the system that no more data follows
immediately, which is equivalent to dropping the handle.

Specifically, in async code the handle can be long-lived and survive await
points, which makes it more important to signal momentary completion to the
system.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>