give_container: Only push if message non-empty
This fixes #523. The reason for the error is that guarded messages are
only emitted when the message is non-empty. In this case however, the
message was empty, recording that a message was sent, but never recording
that the message was received. The fix makes sure to only send within
`give_container` when the message is non-empty. This solves this particular
problem, but the requirements around empty messages are still somewhat
implicit.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>