input handles: give empty containers to operators
The current code tried to be helpful by ignoring any messages with zero
length, but that made it impossible for the caller to differentiate
between "There was an empty container message" vs "There are no more
messages".

In the presence of empty containers in the channels this can lead
operators to prematurely yield back to timely without fully processing
their inputs, which in turn can lead to the whole dataflow not making
progress since timely won't re-activate operators that have leftover
events in their channels.

This PR removes this check and always forwards the containers to the
operators, exactly as they were sent.

Fixes #523

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>