Restructure `TrackerEvent`
This commit changes the shape of `TrackerEvent` to include the
source/target distinction in a `Location` rather than separate enum
variants. Apart from less boilerplate, this speeds up the frontier
logging in `propagate_all` as we can directly use the contents of
`pushed_changes`, instead of having to split them into target and source
changes.

A side effect of this is that the definition of
`ProgressEventTimestampVec` changes, which also affects the
`TimelyProgressEvent` type.