Support GATs and containers in more places
Extract MapInPlace from Map; make Map GAT-aware

This allows us to implement `map` for `TimelyStack`.

Note that this is a breaking change because the `map_in_place` methods
moves from the `Map` trait to the new `MapInPlace` trait. This is required
because it needs mutable access to data, which `TimelyStack` (and others)
cannot provide.