Remove set_external_summary
Remove the `set_external_summary` function because it is not strictly
required and forces the operator to be scheduled during dataflow
construction. This surfaced an inconsistency in probe which would report
incorrect data until scheduled the first time. The fix is to initialize
it correctly and retract the initial information when scheduled the first
time.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>