Define loggers in terms of container builders (#615)
* Define loggers in terms of container builders

The logging infrastructure had some old assumptions built in, such as the
container type and the size of buffers. With this change, we defer to the
container builder pattern to re-use the existing infrastructure. This also
allows us to switch the container type to something else if we'd like to do
so.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Make the communication log event builder public

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Back out some changes, introduce typed logger

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Cleanup

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

---------

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>