Log action can distinguish data from flush (#619)
* Log action can distinguish data from flush

Previously, the log action would receive an empty container when we'd like
to report progress but no new data. We'd only send the empty container when
we didn't have any other data to send. This makes it hard for users to
distinguish between flushing and just data updates. It might be important
to distinguish the two because only on flush we might not be called again
for a while, but otherwise it's very likely that the logger might receive
more data, or sees a flush.

This change alters the signature of the action to accept a `&mut Option<C>`
(where `C` is `CB::Container`), and we pass `Some(container)` on data, and
`None` on flush.

Clients using the logging API need to change their implementation, as both
vectors and `Option` offer a `iter` function, but with different results.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Do not track dirty state in logger

We don't regularly drop the inner logger, so one additional flush doesn't
justify the added complexity.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Align batch logger API closer to logging

`publish_batch` accepts a mutable reference to an option instead of a
mutable reference to a container.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

* Don't clear containers

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>

---------

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>