Avoid allocation in progcaster (#622)
Previously, we'd first clone the changes to be sent downstream before
converting them to a vector. This can cause two allocations, one for
cloning the change batch, and one for the `to_vec` call (it takes self by
reference). This change adjusts it so that we convert the SmallVec into a
vector (`into_vec`).

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>