Flatten reachability logging, log identifier
Previously, reachability and progress logging would use their address to
identify the place where an event happened. In logging this is not great
because every event needs to come with a vector. Instead, use the
operator's global identifier, which we can translate back to the address
using the operates events.

Flattens the reachability logs so they contain no vectors anymore. This
comes at the expense of logging the identifier more frequently. I think
the change is justifiable as the vector previously had at least the size of
three identifiers, plus the memory of the address itself, so only once we
log more than at least three events, we use more space than we did before.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>