Progress container with length
Splits the Container trait into a progress container and a container. The
progress container only exposes a length, which is enough for most parts of
Timely. The container trait extends the progress container and carries the
usual functions to iterate and clear the contents.

I'm open to renaming and moving types, and primarily wanted to check
whether it's possible to split the trait.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>