partition: lazily create and then reuse output sessions
The partition operator was previously unconditionally initializing
sessions for all outputs every time it received an incoming container
even though the container might have contained datums for only a small
subset of the outputs.

This PR makes it so sessions are only lazily initialized when the first
datum for them is observed. Additionally, sessions are reused across
incoming containers as long as the session's capability dominates the
incoming ones.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>