capture: use length prefixed encoding to reliably receive events
The existing capture/replay mechamism would produce a concatenation of
bincode encoded messages. However, on the receiving end the bincode
format does not have enough information for messages to be decoded
accurately from a non-delimited stream, which could lead to the
receiving end getting confused.

This PR fixes the issue by prefixing each messages with its length in
bytes.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>