Retire target and source changes in bulk
If we have many updates per source or target, `propagate_all` will update
an operator's pointstamps as many times as there are updates. Updating
the mutable antichain behind pointstamps can be expensive.

This change extracts the updates from the target and source changes, and
applies them in bulk on the pointstamps in the hope of reducing the cost
of maintaining the antichain.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>