Remove IterContainer (#707)
This change removes the IterContainer type and reworks other types and
implementations to function independently. Notably, this is a breaking
change for users of exotic containers:
* The `Inspect` trait defers to `&C: IntoIterator` to reveal items in
  containers. Not all currently used containers provide this. Vec does,
  but Rc/Arc and Column don't and do not plan to in the future. Users
  should pivot to `inspect_container` instead and internalize the iteration
  in the closure.
* The `DrainContainer` implementaiton for Rc/Arc depends on the wrapped
  type implementing `IntoIterator` for references. This limits where
  wrapped containers that do not provide this can be used. It's mostly
  limited to the core Timely layer that doesn't provide element-by-element
  functionality.

I feel this change is net positive as it defers to the Rust API to iterate
existing types, instead of us providing our own infrastructure. It comes
at a cost of potentially breaking existing code, which seems acceptable.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>